requirements.txt
# Core LangChain and AI dependencies
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.3.0

# Vector Database
qdrant-client>=1.7.0

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.9.0

# Embeddings and Models
sentence-transformers>=2.2.2
tiktoken>=0.5.0

# Web Interface
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0

# API Framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Logging and Monitoring
loguru>=0.7.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# File handling
pathlib2>=2.3.7
watchdog>=3.0.0
